cleared all errors
