+++ src/app/api/auth/signin/route.ts
@@ -1,5 +1,4 @@
 import { NextResponse } from 'next/server';
-import { getSession } from '@/src/lib/session';
 import actotaApi from '@/src/lib/apiClient';
 import axios from 'axios';
 import { setAuthCookie } from '@/src/helpers/auth';
@@ -12,21 +11,23 @@ export async function POST(request: Request) {
       { email: payload.email, password: payload.password },
     );
 
-    if(response.data.auth_token){
+    if (response.data.auth_token) {
 
       setAuthCookie(response.data.auth_token);
       const sessionResponse = await actotaApi.get(
         "/api/auth/session",
-        { headers: {
-          'Authorization': `Bearer ${response.data.auth_token}`,
-        }},
+        {
+          headers: {
+            'Authorization': `Bearer ${response.data.auth_token}`,
+          }
+        },
       );
       return NextResponse.json(
         { success: true, message: 'Login successful', data: sessionResponse.data },
         { status: 200 }
       );
     }
-  } catch (error:any) {
+  } catch (error: any) {
     return NextResponse.json(
       { success: false, message: error.message || 'An error occurred' },
       { status: 500 }
