+++ src/components/FeaturedItineraries.tsx
@@ -14,9 +14,10 @@ export default function FeaturedItineraries() {
   const [isLoading, setIsLoading] = React.useState(true);
   const [selectedActivity, setSelectedActivity] = useState([]);
 
+
   useEffect(() => {
     if (itineraries) {
-      const filteredListings = itineraries?.map((listing: any) => favorites?.some((favorite: any) => favorite._id.$oid === listing._id.$oid) ? { ...listing, isFavorite: true } : listing);
+      const filteredListings = itineraries?.data.map((listing: any) => favorites?.some((favorite: any) => favorite._id.$oid === listing._id.$oid) ? { ...listing, isFavorite: true } : listing);
       setListings(filteredListings);
     }
   }, [itineraries]);
